Update about.md
I have rewritten the text of this page to be as clear and structured as I could, per the discussions in the UX meeting with @vabishaa. 

In particular, I have tried to ensure internal consistency within the document in the use/spelling of the terms: "VP-Portal" (taken from the first heading) and "Virtual Platform". 

I have removed the word "Network" where it appeared as it was inconsistently used, and added to a lack of clarity. I have also tried to avoid the use of the acronym "VP" when referring to the Virtual Platform. 

I don't know if there is already a standardised style guide for all of these terms, but in the absence of that, I hope this is OK.